# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GameDash API
  version: 0.1.0
  description: Baseline contract for GameDash MVP foundation.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      security: []
      operationId: register
      tags: [Auth]
      summary: Register a player account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/login:
    post:
      security: []
      operationId: login
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/refresh:
    post:
      security: []
      operationId: refreshToken
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      operationId: getCurrentAuthUser
      tags: [Auth]
      summary: Get authenticated user and player profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: Logout and revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /maps:
    get:
      operationId: listMaps
      tags: [Maps]
      summary: Browse maps
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MapStatus"
        - name: creatorId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Map list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createMap
      tags: [Maps]
      summary: Create map
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMapRequest"
      responses:
        "201":
          description: Map created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /maps/creators/{creatorId}/stats:
    get:
      operationId: getCreatorMapStats
      tags: [Maps]
      summary: Get creator map statistics
      parameters:
        - name: creatorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Creator map statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorMapStatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /maps/{mapId}:
    get:
      operationId: getMap
      tags: [Maps]
      summary: Get map detail
      parameters:
        - $ref: "#/components/parameters/mapId"
      responses:
        "200":
          description: Map detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /maps/{mapId}/stats:
    get:
      operationId: getMapStats
      tags: [Maps]
      summary: Get map statistics
      parameters:
        - $ref: "#/components/parameters/mapId"
      responses:
        "200":
          description: Map statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapStatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /maps/{mapId}/versions:
    post:
      operationId: createMapVersion
      tags: [Maps]
      summary: Create map version
      parameters:
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMapVersionRequest"
      responses:
        "201":
          description: Version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapVersionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /maps/{mapId}/votes:
    post:
      operationId: voteMap
      tags: [Maps]
      summary: Vote a map
      parameters:
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteMapRequest"
      responses:
        "201":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapInteractionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /maps/{mapId}/tests:
    post:
      operationId: testMap
      tags: [Maps]
      summary: Record map test
      parameters:
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestMapRequest"
      responses:
        "201":
          description: Test recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapInteractionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /maps/{mapId}/favorites:
    post:
      operationId: favoriteMap
      tags: [Maps]
      summary: Favorite or unfavorite a map
      parameters:
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteMapRequest"
      responses:
        "201":
          description: Favorite state recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapInteractionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [Admin]
      summary: Get studio dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboardSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/settings:
    get:
      operationId: getStudioSettings
      tags: [Admin]
      summary: Get studio tuning settings
      responses:
        "200":
          description: Studio settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioSettingsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      operationId: updateStudioSettings
      tags: [Admin]
      summary: Update studio tuning settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudioSettingsRequest"
      responses:
        "200":
          description: Updated studio settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioSettingsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/moderation/signals:
    get:
      operationId: getModerationSignals
      tags: [Admin]
      summary: Get moderation signals and reports
      responses:
        "200":
          description: Moderation signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModerationSignalResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/moderation/history:
    get:
      operationId: getModerationHistory
      tags: [Admin]
      summary: Get moderation action history
      responses:
        "200":
          description: Moderation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModerationActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/moderation/accounts/{userId}/actions:
    post:
      operationId: moderateAccount
      tags: [Admin]
      summary: Apply account moderation action
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountModerationRequest"
      responses:
        "201":
          description: Account moderation action recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/moderation/maps/{mapId}/actions:
    post:
      operationId: moderateMap
      tags: [Admin]
      summary: Apply map moderation action
      parameters:
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapModerationRequest"
      responses:
        "201":
          description: Map moderation action recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    mapId:
      name: mapId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated user does not have the required role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Role:
      type: string
      enum: [player, staff, admin]
...
    MapStatus:
      type: string
      enum: [draft, beta, stable, hidden]
    RegisterRequest:
      type: object
      required: [email, password, pseudo]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        pseudo:
          type: string
        avatarUrl:
          type: string
          format: uri
        region:
          type: string
        bio:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    PlayerProfileResponse:
      type: object
      required: [userId, pseudo]
      properties:
        userId:
          type: string
        pseudo:
          type: string
        avatarUrl:
          type: string
          format: uri
        region:
          type: string
        bio:
          type: string
...
    AuthUserResponse:
      type: object
      required: [id, email, role, profile]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        profile:
          $ref: "#/components/schemas/PlayerProfileResponse"
    AuthTokensResponse:
      type: object
      required: [accessToken, refreshToken, role, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        user:
          $ref: "#/components/schemas/AuthUserResponse"
...
    CreateMapRequest:
      type: object
      required: [title, description, status]
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/MapStatus"
    CreateMapVersionRequest:
      type: object
      required: [versionLabel, releaseNotes]
      properties:
        versionLabel:
          type: string
          example: v1
        releaseNotes:
          type: string
    MapVersionResponse:
      type: object
      required: [id, mapId, versionLabel, releaseNotes, createdAt]
      properties:
        id:
          type: string
          format: uuid
        mapId:
          type: string
          format: uuid
        versionLabel:
          type: string
        releaseNotes:
          type: string
        createdAt:
          type: string
          format: date-time
    VoteMapRequest:
      type: object
      required: [value]
      properties:
        value:
          type: integer
          enum: [-1, 1]
    TestMapRequest:
      type: object
      required: [completed]
      properties:
        completed:
          type: boolean
    FavoriteMapRequest:
      type: object
      required: [favorited]
      properties:
        favorited:
          type: boolean
    ActionAcceptedResponse:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean
    MapStatsResponse:
      type: object
      required:
        [
          mapId,
          versionCount,
          voteScore,
          upvotes,
          downvotes,
          completedTests,
          favorites,
          popularityScore
        ]
      properties:
        mapId:
          type: string
          format: uuid
        versionCount:
          type: integer
          minimum: 0
        voteScore:
          type: integer
        upvotes:
          type: integer
          minimum: 0
        downvotes:
          type: integer
          minimum: 0
        completedTests:
          type: integer
          minimum: 0
        favorites:
          type: integer
          minimum: 0
        popularityScore:
          type: number
          format: float
    MapSummary:
      type: object
      required:
        [
          id,
          creatorId,
          title,
          description,
          tags,
          status,
          popularityScore,
          createdAt,
          updatedAt,
          stats
        ]
      properties:
        id:
          type: string
          format: uuid
        creatorId:
          type: string
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/MapStatus"
        popularityScore:
          type: number
          format: float
        latestVersionLabel:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        stats:
          $ref: "#/components/schemas/MapStatsResponse"
    MapDetailResponse:
      allOf:
        - $ref: "#/components/schemas/MapSummary"
        - type: object
          required: [versions]
          properties:
            versions:
              type: array
              items:
                $ref: "#/components/schemas/MapVersionResponse"
    MapInteractionResponse:
      allOf:
        - $ref: "#/components/schemas/ActionAcceptedResponse"
        - type: object
          required: [map]
          properties:
            map:
              $ref: "#/components/schemas/MapSummary"
    CreatorMapStatsResponse:
      type: object
      required:
        [
          creatorId,
          totalMaps,
          publishedMaps,
          totalVotes,
          totalTests,
          totalFavorites,
          averagePopularityScore
        ]
      properties:
        creatorId:
          type: string
        totalMaps:
          type: integer
          minimum: 0
        publishedMaps:
          type: integer
          minimum: 0
        totalVotes:
          type: integer
          minimum: 0
        totalTests:
          type: integer
          minimum: 0
        totalFavorites:
          type: integer
          minimum: 0
        averagePopularityScore:
          type: number
          format: float
    AdminDashboardSummary:
      type: object
      required:
        [
          activePlayers,
          dailyMatches,
          virtualRevenue,
          mapActivity,
          openModerationSignals,
          activeSanctions,
          settingsLastUpdated
        ]
      properties:
        activePlayers:
          type: integer
        dailyMatches:
          type: integer
        virtualRevenue:
          type: integer
        mapActivity:
          type: integer
        openModerationSignals:
          type: integer
          minimum: 0
        activeSanctions:
          type: integer
          minimum: 0
        settingsLastUpdated:
          type: string
          format: date-time
    MatchmakingSettings:
      type: object
      required: [rankedQueueMaxWaitSeconds, funQueueMaxWaitSeconds, matchSize]
      properties:
        rankedQueueMaxWaitSeconds:
          type: integer
          minimum: 0
        funQueueMaxWaitSeconds:
          type: integer
          minimum: 0
        matchSize:
          type: integer
          minimum: 2
    MmrSettings:
      type: object
      required: [placementMmr, rankedWinDelta, rankedLossDelta, unrankedWinDelta, unrankedLossDelta]
      properties:
        placementMmr:
          type: integer
          minimum: 0
        rankedWinDelta:
          type: integer
        rankedLossDelta:
          type: integer
        unrankedWinDelta:
          type: integer
        unrankedLossDelta:
          type: integer
    EconomySettings:
      type: object
      required: [starterSoftBalance, starterHardBalance, purchaseEnabled, refundWindowHours]
      properties:
        starterSoftBalance:
          type: integer
          minimum: 0
        starterHardBalance:
          type: integer
          minimum: 0
        purchaseEnabled:
          type: boolean
        refundWindowHours:
          type: integer
          minimum: 0
    StudioSettingsResponse:
      type: object
      required: [matchmaking, mmr, economy, updatedAt]
      properties:
        matchmaking:
          $ref: "#/components/schemas/MatchmakingSettings"
        mmr:
          $ref: "#/components/schemas/MmrSettings"
        economy:
          $ref: "#/components/schemas/EconomySettings"
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    UpdateStudioSettingsRequest:
      type: object
      properties:
        matchmaking:
          type: object
          properties:
            rankedQueueMaxWaitSeconds:
              type: integer
              minimum: 0
            funQueueMaxWaitSeconds:
              type: integer
              minimum: 0
            matchSize:
              type: integer
              minimum: 2
        mmr:
          type: object
          properties:
            placementMmr:
              type: integer
              minimum: 0
            rankedWinDelta:
              type: integer
            rankedLossDelta:
              type: integer
            unrankedWinDelta:
              type: integer
            unrankedLossDelta:
              type: integer
        economy:
          type: object
          properties:
            starterSoftBalance:
              type: integer
              minimum: 0
            starterHardBalance:
              type: integer
              minimum: 0
            purchaseEnabled:
              type: boolean
            refundWindowHours:
              type: integer
              minimum: 0
    AccountModerationAction:
      type: string
      enum: [warn, suspend, ban]
    MapModerationAction:
      type: string
      enum: [hide, restore, feature]
    ModerationTargetType:
      type: string
      enum: [account, map]
    ModerationSignalStatus:
      type: string
      enum: [open, reviewed, dismissed]
    AccountModerationRequest:
      type: object
      required: [action, reason]
      properties:
        action:
          $ref: "#/components/schemas/AccountModerationAction"
        reason:
          type: string
        durationHours:
          type: integer
          minimum: 1
    MapModerationRequest:
      type: object
      required: [action, reason]
      properties:
        action:
          $ref: "#/components/schemas/MapModerationAction"
        reason:
          type: string
    ModerationActionResponse:
      type: object
      required: [id, targetType, targetId, action, reason, actorId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        targetType:
          $ref: "#/components/schemas/ModerationTargetType"
        targetId:
          type: string
        action:
          type: string
        reason:
          type: string
        actorId:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    ModerationSignalResponse:
      type: object
      required: [id, targetType, targetId, reason, status, source, createdAt]
      properties:
        id:
          type: string
        targetType:
          $ref: "#/components/schemas/ModerationTargetType"
        targetId:
          type: string
        reason:
          type: string
        status:
          $ref: "#/components/schemas/ModerationSignalStatus"
        source:
          type: string
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, statusCode, timestamp, path, requestId]
          properties:
            code:
              type: string
              example: bad_request
            message:
              type: string
            statusCode:
              type: integer
            timestamp:
              type: string
              format: date-time
            path:
              type: string
              example: /api/v1/auth/login
            requestId:
              type: string